backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-genai==0.1.0
